test create users endpoint
