+++ index.js
@@ -1,22 +1,7 @@
-import express from 'express'
-import 'regenerator-runtime/runtime'
-import './src/db/mongoose'
-import userRouter from './routers/user'
-import taskRouter from './routers/task'
-import cors from 'cors'
-const app = express()
+import app from './app'
+
 const port = process.env.PORT
 
-app.use(express.json())
-app.use(cors())
-app.use([userRouter, taskRouter])
-
-app.get('/', async (req, res) => {
-    res.status(200).send({
-        status: 200,
-        data: 'Welcome!!!'
-    })
-})
 app.listen(port, () => {
     console.log(`Runing on port ${port}`)
